python-service/requirements.txt
Flask==3.0.3
pdf2image==1.17.0
Pillow==10.4.0
